+++ src/textAngular-sanitize.js
@@ -548,15 +548,37 @@ function validStyles(styleAttr){
           value === 'italic'
         )
       ||
-				key === 'float' && (
-					value === 'left'
-					|| value === 'right'
-					|| value === 'none'
-				)
+        key === 'float' && (
+            value === 'left'
+            || value === 'right'
+            || value === 'none'
+        )
+      ||
+        key === 'vertical-align' && (
+            value === 'baseline'
+            || value === 'sub'
+            || value === 'super'
+            || value === 'test-top'
+            || value === 'text-bottom'
+            || value === 'middle'
+            || value === 'top'
+            || value === 'bottom'
+            || value.match(/[0-9]*(px|em)/)
+            || value.match(/[0-9]+?%/)
+        )
       ||
         key === 'font-size' && (
-					value.match(/[0-9\.]*(px|em|rem|%)/)
-				)
+            value === 'xx-small'
+            || value === 'x-small'
+            || value === 'small'
+            || value === 'medium'
+            || value === 'large'
+            || value === 'x-large'
+            || value === 'xx-large'
+            || value === 'larger'
+            || value === 'smaller'
+            || value.match(/[0-9]*\.?[0-9]*(px|em|rem|mm|q|cm|in|pt|pc|%)/)
+                               )
 			||
 				(key === 'width' || key === 'height') && (
 					value.match(/[0-9\.]*(px|em|rem|%)/)